+++ pkg/orders/v2/v2.go
@@ -1,4 +1,4 @@
-package orders
+package v2
 
 import (
 	"context"
